requirements.txt
discord.py
requests
redis
pyro4
python-twitter